% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocs.R
\name{build_summary_tocs}
\alias{build_summary_tocs}
\title{Build Totals and Prorated Totals for Full Test and Subdomains}
\usage{
build_summary_tocs(
  df = NULL,
  max_missing = NULL,
  age_var = "age",
  gender_var = "gender",
  respondent_var = "p_respondent",
  required_test_cols = NULL
)
}
\arguments{
\item{df}{should be a data.frame from \code{\link[=clean_file]{clean_file()}}}

\item{max_missing}{max_missing is passed from the \code{\link[=score_tocs2]{score_tocs2()}} function. By default, the tocs allows 0 missing items.}

\item{age_var}{Name of the age variable in your data}

\item{gender_var}{Name of the gender variable in your data}

\item{respondent_var}{Name of the respondent variable in your data}

\item{required_test_cols}{An array of the names of the questionnaires questions, i.e. swan1, swan2, etc}
}
\value{
A data frame with all of the totals columns
}
\description{
Use the dataframe from \code{\link[=clean_file]{clean_file()}} and the \code{\link[=mkpro]{mkpro()}} function to reverse scores, then
calculate totals, missingness, and pro-rated totals for the total test and subdomains
}
