% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{get_swan_tscores}
\alias{get_swan_tscores}
\title{Run analysis on SWAN raw values to return t-scores}
\usage{
get_swan_tscores(file = NULL, output_folder = here::here())
}
\arguments{
\item{file}{Pathway to formatted raw SWAN scores. If left blank file finder will pop up to allow you to select the file.}

\item{output_folder}{Output file pathway
\enumerate{
\item Leave blank - This will output a csv file with the t-scores to your working directory
\item Specify a pathway - This will output a csv file to the specified pathway
\item Set to `NULL` - This will not output a csv file
}}
}
\value{
table with t-scores attached to raw swan values
}
\description{
get_swan_tscores() returns gendered and non-gendered t-scores for the Strengths and Weaknesses of ADHD Symptoms and Normal Behavior Rating Scale (SWAN) assessment
}
