% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocs.R
\name{get_tocs_tscores}
\alias{get_tocs_tscores}
\title{Run analysis on TOCS raw values to return t-scores}
\usage{
get_tocs_tscores(file = NULL, output_folder = here::here())
}
\arguments{
\item{file}{Pathway to formatted raw scores. If left blank file finder will pop up to allow you to select the file.}

\item{output_folder}{Output file pathway
\enumerate{
\item Leave blank - This will output a csv file with the t-scores to your working directory
\item Specify a pathway - This will output a csv file to the specified pathway
\item Set to `NULL` - This will not output a csv file
}}
}
\value{
table with t-scores attached to raw swan values
}
\description{
get_tocs_tscores() returns gendered and non-gendered t-scores for the [Toronto Obsessive-Compulsive Scale (TOCS) assessment](https://pubmed.ncbi.nlm.nih.gov/27015722/)
}
