% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{mkvars}
\alias{mkvars}
\title{Make Variables - Subset SWAN to subdomains.}
\usage{
mkvars(a = NULL, b = NULL, root = "swan")
}
\arguments{
\item{a}{First question of subset}

\item{b}{Last question of subset}

\item{root}{Root name of}
}
\value{
A data frame ready for use or an error
}
\description{
Pass the root of the test with the question numbers to subset the SWAN. 1-9 = Inattentive. 10-18 = Hyperactive.
Function to list all questionnaire items (and not have to type them out) - used throughout
a:  first item number (usually "1", but when referencing subdomains, or for SWAN ODD, first item may be something other than 1
b:  last item number
root:  part of the item name that doesn't change (eg:  for swan1 to swan18, the root is "swan" )
}
\author{
Annie
}
