% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{run_model_old_swan}
\alias{run_model_old_swan}
\title{An old version of the model, initially written in Annie's language}
\usage{
run_model_old_swan(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from \code{\link[=build_summary_swan]{build_summary_swan()}}}
}
\value{
A data frame with t-scores
}
\description{
This is the generic model used to create t-scores. It adjusts for gender, age, respondent, and time.
It is best used in all cases unless a team is trying to look at values over time
Use the dataframe from \code{\link[=build_summary_swan]{build_summary_swan()}} to produce t-scores
}
