% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{run_model_swan}
\alias{run_model_swan}
\title{Runs the model that adjusts for time and create t-scores}
\usage{
run_model_swan(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from [build_summary()]}
}
\value{
A data frame with t-scores
}
\description{
This is the generic model used to create t-scores. It adjusts for gender, age, respondent, and time.
It is best used in all cases unless a team is trying to look at values over time
Use the dataframe from [build_summary()] to produce t-scores
}
