% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocs.R
\name{run_model_tocs}
\alias{run_model_tocs}
\title{Runs the model that adjusts for time and create t-scores}
\usage{
run_model_tocs(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from \code{\link[=build_summary_tocs]{build_summary_tocs()}}}
}
\value{
A data frame with t-scores
}
\description{
This is the generic model used to create t-scores. It adjusts for gender, age, respondent, and time.
It is best used in all cases unless a team is trying to look at values over time
Use the dataframe from \code{\link[=build_summary_tocs]{build_summary_tocs()}} to produce t-scores
}
