% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocs.R
\name{score_tocs2}
\alias{score_tocs2}
\title{Run analysis on TOCS raw values to return t-scores}
\usage{
score_tocs2(
  df = NULL,
  file = FALSE,
  output_folder = NULL,
  age_var = "age",
  gender_var = "gender",
  respondent_var = "p_respondent",
  tocs2_vars = paste0("tocs", seq(1, 24)),
  max_missing = 0,
  ignore_check = FALSE
)
}
\arguments{
\item{df}{If you already have the TOCS-2 data in your R environment, pass the dataframe to this parameter}

\item{file}{If you prefer scoring a spreadsheet...
\enumerate{
\item Change to \code{TRUE} to pop-up a finder to allow you select a file. Alternatively, leave df and file empty to pop-up a finder.
\item Or specify a pathway
}}

\item{output_folder}{Optional, output file pathway. Defauts to \code{NULL}. Specify a pathway to output a csv file.}

\item{age_var}{Name of the age variable in your data}

\item{gender_var}{Name of the gender variable in your data}

\item{respondent_var}{Name of the respondent variable in your data}

\item{tocs2_vars}{Column names of the 24 TOCS questions}

\item{max_missing}{By default, 0 items are allowed to be missing on the TOCS. Any questionnaire with 1 or more missing, will not be scored. If you'd like to adjust this number, change the max_missing value.
This will use a prorated score to generate t-scores. Please be aware that missingness can induce issues when analyzing.}

\item{ignore_check}{Data are validated to look for missing or improperly formatted values before scoring. Errors are thrown when data aren't valid; however, this can cause issues
in real data sets where data vary for good reasons. To skip the validation process, set ignore_check to \code{TRUE}. NAs will be returned where data are invalid}
}
\value{
A dataframe where... \cr
\itemize{
\item T-scores and total scores columns are added
\item SWAN questions are reversed scored so that higher numbers match increased symptoms.
\item Otherwise all columns are not modified
}
}
\description{
score_tocs2() returns gendered and non-gendered t-scores for the \href{https://pubmed.ncbi.nlm.nih.gov/27015722/}{Toronto Obsessive-Compulsive Scale (TOCS) assessment}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

# Read in the file of scores
# This is an example file
csv <- system.file("extdata", "sample_tocs.csv", package = "sfsScorer")

# Score via the file parameter
scores_csv <- score_tocs2(file = csv)

# Score via the df paramter
df <- rio::import(csv)
scores_csv <- score_tocs2(df = df)

# The data are automatically validated.
# To ignore the validation errors and introduce `NA`, set `ignore_check = TRUE`
df_mod <- df |>
  dplyr::mutate(p_respondent = 2)
scores_csv <- score_tocs2(df = df_mod, ignore_check = TRUE)

}
