% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfsScorer-package.R
\docType{package}
\name{sfsScorer-package}
\alias{sfsScorer}
\alias{sfsScorer-package}
\title{sfsScorer: Scores child psychometric questionnaires using norms from the Spit for Science study}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Spit for Science (SFS) was a cross-sectional mental health study performed by Dr. Crosbie and Dr. Schachar, in which over 40,000 child and youth participants contributed genetic and questionnaire data. SFS is one of the largest studies to examine child and youth mental health, and as such, it has been used to validate multiple psychometric questionnaires. This package scores and produces t-scoresfor the SFS questionnaires, i.e. the Strengths and Weaknesses of ADHD Symptoms and Normal Behavior Rating Scale (SWAN), the Toronto Obsessive-Compulsive Scale (TOCS), etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jclutton/sfsScorer}
  \item \url{https://jclutton.github.io/sfsScorer/}
  \item Report bugs at \url{https://github.com/jclutton/sfsScorer/issues}
}

}
\author{
\strong{Maintainer}: Jonathan Clutton \email{jonathan.clutton@sickkids.ca} (\href{https://orcid.org/0000-0002-3593-4271}{ORCID})

}
\keyword{internal}
