% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{build_summary_swan}
\alias{build_summary_swan}
\title{Build Totals and Prorated Totals for Full Test and Subdomains}
\usage{
build_summary_swan(df = NULL)
}
\arguments{
\item{df}{should be a data.frame from [clean_file()]}
}
\value{
A data frame with all of the totals columns
}
\description{
Use the dataframe from [clean_file()] and the [mkpro()] function to reverse scores, then
calculate totals, missingness, and pro-rated totals for the total test and subdomains
}
