% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tocs.R
\name{build_summary_tocs}
\alias{build_summary_tocs}
\title{Build Totals and Prorated Totals for Full Test and Subdomains}
\usage{
build_summary_tocs(df = NULL, max_missing = NULL)
}
\arguments{
\item{df}{should be a data.frame from \code{\link[=clean_file]{clean_file()}}}

\item{max_missing}{max_missing is passed from the \code{\link[=get_tocs_tscores]{get_tocs_tscores()}} function. By default, the tocs allows 0 missing items.}
}
\value{
A data frame with all of the totals columns
}
\description{
Use the dataframe from \code{\link[=clean_file]{clean_file()}} and the \code{\link[=mkpro]{mkpro()}} function to reverse scores, then
calculate totals, missingness, and pro-rated totals for the total test and subdomains
}
