% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{clean_file}
\alias{clean_file}
\title{Clean File}
\usage{
clean_file(df = NULL, test = NULL)
}
\arguments{
\item{df}{The df function allows you to point to a dataframe as opposed to a file}

\item{test}{Which questionnaire are we running}
}
\value{
A clean data frame ready for t-scores
}
\description{
This function runs checks to be sure that the file and non-test specific data are formatted correctly
}
