% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{clean_file}
\alias{clean_file}
\title{Clean File}
\usage{
clean_file(file_path = NULL)
}
\arguments{
\item{file_path}{Should be a path on your computer to the SWAN scores}
}
\value{
A clean data frame ready for t-scores
}
\description{
This function checks that the uploaded SWAN score file does not have any issues with the SWAN
values or the gender and age coding
}
