% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{mkpro}
\alias{mkpro}
\title{Make Prorated Scores}
\usage{
mkpro(
  maxmiss = NA,
  dat = NA,
  a = NULL,
  b = NULL,
  root = "swan",
  newroot = "swan"
)
}
\arguments{
\item{maxmiss}{maximum number of missing values before can be considered invalid}

\item{dat}{should be a data.frame from \code{\link[=clean_file]{clean_file()}}}

\item{a}{First question of subset}

\item{b}{Last question of subset}

\item{root}{Root name of}

\item{newroot}{a new name if root names need to be changed}
}
\value{
A data frame ready for use or an error
}
\description{
a: first item number in questionnaire (usually "1")
}
\author{
Annie
}
