% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{score_swan}
\alias{score_swan}
\title{Run analysis on SWAN raw values to return t-scores}
\usage{
score_swan(df = NULL, file = FALSE, output_folder = NULL, ignore_check = FALSE)
}
\arguments{
\item{df}{If you already have the SWAN data in your R environment, pass the dataframe to this parameter}

\item{file}{If you prefer scoring a spreadsheet...
\enumerate{
\item Change to \code{TRUE} to pop-up a finder to allow you select a file. Alternatively, leave df and file empty to pop-up a finder.
\item Or specify a pathway
}}

\item{output_folder}{Optional, output file pathway. Defauts to \code{NULL}. Specify a pathway to output a csv file.}

\item{ignore_check}{Data are validated to look for missing or improperly formatted values before scoring. Errors are thrown when data aren't valid; however, this can cause issues
in real data sets where data vary for good reasons. To skip the validation process, set ignore_check to \code{TRUE}. NAs will be returned where data are invalid}
}
\value{
table with t-scores attached to raw swan values
}
\description{
score_swan() returns gendered and non-gendered t-scores for the Strengths and Weaknesses of ADHD Symptoms and Normal Behavior Rating Scale (SWAN) assessment

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Read in the file of scores
# This is an example file
csv <- system.file("extdata", "sample_swan.csv", package = "sfsScorer")

# Score via the file parameter
scores_csv <- score_swan(file = csv)

# Score via the df paramter
df <- rio::import(csv)
scores_csv <- score_swan(df = df)

# The data are automatically validated.
# To ignore the validation errors and introduce `NA`, set `ignore_check = TRUE`
df_mod <- df |>
  dplyr::mutate(swan1 = 6)
scores_csv <- score_swan(df = df, ignore_check = TRUE)


}
