% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swan.R
\name{score_swan}
\alias{score_swan}
\title{Run analysis on SWAN raw values to return t-scores}
\usage{
score_swan(df = NULL, file = FALSE, output_folder = NULL, max_missing = 0)
}
\arguments{
\item{df}{If you already have the TOCS-2 data in your R environment, pass the dataframe to this parameter}

\item{file}{If you prefer scoring a spreadsheet...
\enumerate{
\item TRUE - This will pop-up a finder to allow you select a file
\item Specify a pathway
}}

\item{output_folder}{Output file pathway
\enumerate{
\item Leave blank - This will output a csv file with the t-scores to your working directory
\item Specify a pathway - This will output a csv file to the specified pathway
\item Set to \code{NULL} - This will not output a csv file
}}

\item{max_missing}{By default, 0 items are allowed to be missing on the TOCS. Any questionnaire with 1 or more missing, will not be scored. If you'd like to adjust this number, change the max_missing value.
This will use a prorated score to generate t-scores. Please be aware that missingness can induce issues when analyzing.}
}
\value{
table with t-scores attached to raw swan values
}
\description{
score_swan() returns gendered and non-gendered t-scores for the Strengths and Weaknesses of ADHD Symptoms and Normal Behavior Rating Scale (SWAN) assessment

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
